requirements.txt
# Core dependencies
python-dotenv>=1.0.0
pydantic>=2.0.0
typing-extensions>=4.6.0

# Audio processing
pyaudio>=0.2.13
numpy>=1.24.0
wave>=0.0.2

# API and server
fastapi>=0.100.0
uvicorn>=0.22.0
python-multipart>=0.0.6

# WebSocket client
aiohttp>=3.8.5
websockets>=11.0.3

# Deepgram SDK
deepgram-sdk>=2.11.0

# Testing
pytest>=7.3.1
pytest-asyncio>=0.21.0

# Demo script dependencies
argparse>=1.4.0
pathlib>=1.0.1 